+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- golang commands were wrong
+- there was a output on report subcommands breaking the format
+
 ## [v0.27.1] - 2021-12-31
 
 ### Fixed