+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `log` subcommand deprecated in favor of `report`
+- default `report` now allows to set only one date of the range, in this situation it will treat
+  start and end date as being the same.
+
 ### Added
 
 - new `report today` to show only the time entries from today, with `report` options