+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.29.0] - 2022-01-12
+
 ### Changed
 
 - `show` subcommand has its parameter as optional, and shows current time entry by default when the
@@ -768,8 +770,9 @@ time entry.
 - Golang CLI using [cobra](https://github.com/spf13/cobra)
 - Makefile to help setup actions
 
-[Unreleased]: https://github.com/lucassabreu/clockify-cli/compare/v0.28.0...HEAD
-[v0.28.0]: https://github.com/lucassabreu/clockify-cli/releases/tag/v0.28.1
+[Unreleased]: https://github.com/lucassabreu/clockify-cli/compare/v0.29.0...HEAD
+[v0.29.0]: https://github.com/lucassabreu/clockify-cli/releases/tag/v0.29.0
+[v0.28.0]: https://github.com/lucassabreu/clockify-cli/releases/tag/v0.28.0
 [v0.27.1]: https://github.com/lucassabreu/clockify-cli/releases/tag/v0.27.1
 [v0.27.0]: https://github.com/lucassabreu/clockify-cli/releases/tag/v0.27.0
 [v0.26.1]: https://github.com/lucassabreu/clockify-cli/releases/tag/v0.26.1